Moved util fns to DeadlineSelector/utils.ts
