+++ src/cow-react/modules/limitOrders/pure/DeadlineSelector/index.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react'
 import { Dropdown } from '@cow/common/pure/Dropdown'
-import { LimitOrderDeadline, limitOrdersDeadlines, MAX_CUSTOM_DEADLINE, MIN_CUSTOM_DEADLINE } from './deadlines'
+import { LimitOrderDeadline, limitOrdersDeadlines } from './deadlines'
 import { GpModal as Modal } from '@src/custom/components/Modal'
 
 import { useCallback, useMemo, useRef } from 'react'
@@ -21,6 +21,12 @@ import {
 } from './styled'
 import { Trans } from '@lingui/macro'
 import { ButtonPrimary, ButtonSecondary } from '@src/components/Button'
+import {
+  calculateMinMax,
+  formatDateToLocalTime,
+  getTimeZoneOffset,
+  limitDateString,
+} from '@cow/modules/limitOrders/pure/DeadlineSelector/utils'
 
 const CUSTOM_DATE_OPTIONS: Intl.DateTimeFormatOptions = {
   year: '2-digit',
@@ -42,10 +48,10 @@ export function DeadlineSelector(props: DeadlineSelectorProps) {
   const { deadline, customDeadline, selectDeadline, selectCustomDeadline } = props
 
   const currentDeadlineNode = useRef<HTMLButtonElement | null>(null)
-  const [[minDate, maxDate], setMinMax] = useState<[Date, Date]>(_calculateMinMax)
+  const [[minDate, maxDate], setMinMax] = useState<[Date, Date]>(calculateMinMax)
 
-  const min = _limitDateString(minDate)
-  const max = _limitDateString(maxDate)
+  const min = limitDateString(minDate)
+  const max = limitDateString(maxDate)
 
   const [error, setError] = useState<string | null>(null)
   const [value, setValue] = useState<string>('')
@@ -88,11 +94,11 @@ export function DeadlineSelector(props: DeadlineSelectorProps) {
 
   // Sets value from input, if it exists
   const onChange = useCallback(
-    ({ target: { value, valueAsNumber } }: React.ChangeEvent<HTMLInputElement>) => {
+    ({ target: { value } }: React.ChangeEvent<HTMLInputElement>) => {
       // Some browsers offer a `clear` button in their date picker
       // That action sets the value to `''`
       // In that case, use the default min value
-      setValue(value || _formatDateToLocalTime(minDate))
+      setValue(value || formatDateToLocalTime(minDate))
     },
     [minDate]
   )
@@ -102,15 +108,15 @@ export function DeadlineSelector(props: DeadlineSelectorProps) {
     currentDeadlineNode.current?.click() // Close dropdown
     setIsOpen(true)
     setError(null)
-    setMinMax(_calculateMinMax()) // Update min/max every time modal is open
-    setValue(_formatDateToLocalTime(customDeadline || minDate)) // reset input to clear unsaved values
+    setMinMax(calculateMinMax()) // Update min/max every time modal is open
+    setValue(formatDateToLocalTime(customDeadline || minDate)) // reset input to clear unsaved values
   }
   const onDismiss = useCallback(() => setIsOpen(false), [])
 
   const setCustomDeadline = useCallback(() => {
     // `value` is a timezone aware string
     // thus, we append the timezone offset (if any) when building the date object
-    const newDeadline = Math.round(new Date(value + _getTimeZoneOffset()).getTime() / 1000)
+    const newDeadline = Math.round(new Date(value + getTimeZoneOffset()).getTime() / 1000)
 
     selectCustomDeadline(newDeadline)
     onDismiss()
@@ -180,69 +186,4 @@ export function DeadlineSelector(props: DeadlineSelectorProps) {
       </Modal>
     </Wrapper>
   )
-}
-
-function _limitDateString(date: Date | number): string {
-  const _date = typeof date === 'number' ? new Date(date * 1000) : date
-
-  const [first, second] = _date.toISOString().split(':')
-
-  return [first, second].join(':')
-}
-
-const LOCAL_DATE_FORMATTER = new Intl.DateTimeFormat(
-  'en-CA', // using CA because the format is yyyy-mm-dd, which is what we need
-  {
-    year: 'numeric',
-    month: '2-digit',
-    day: '2-digit',
-    hour: '2-digit',
-    minute: '2-digit',
-    hour12: false,
-  }
-)
-
-/**
- * Formats a date object into a Timezone aware string in the format: `yyyy-mm-ddTHH:MM`
- *
- * The important part is the timezone awareness.
- * The `datetime-local` input can't deal with timezones while all Date objs are tz aware.
- * When passing the initial value to the input, using the ISO string representation of the date
- * causes it to be in a different time from user's
- */
-function _formatDateToLocalTime(date: Date | number): string {
-  const _date = typeof date === 'number' ? new Date(date * 1000) : date
-
-  return LOCAL_DATE_FORMATTER.format(_date) // this returns `2017-06-01, 08:30`
-    .replace(/, /, 'T') // we want `2017-06-01T08:30`
-}
-
-function _calculateMinMax(): [Date, Date] {
-  const now = Date.now()
-  return [_trimSeconds(new Date(now + MIN_CUSTOM_DEADLINE)), _trimSeconds(new Date(now + MAX_CUSTOM_DEADLINE))]
-}
-
-function _trimSeconds(date: Date): Date {
-  return new Date(new Date(date.setMilliseconds(0)).setSeconds(0))
-}
-
-/**
- * Get client side timezone offset
- *
- * @returns {(+|-)HH:mm} - Where `HH` is 2 digits hours and `mm` 2 digits minutes.
- *
- * From https://stackoverflow.com/a/30377368/1272513 on 2023/01/31
- */
-function _getTimeZoneOffset() {
-  const timezoneOffset = new Date().getTimezoneOffset()
-  const offset = Math.abs(timezoneOffset)
-  const offsetOperator = timezoneOffset < 0 ? '+' : '-'
-  const offsetHours = Math.floor(offset / 60)
-    .toString()
-    .padStart(2, '0')
-  const offsetMinutes = Math.floor(offset % 60)
-    .toString()
-    .padStart(2, '0')
-
-  return `${offsetOperator}${offsetHours}:${offsetMinutes}`
 }
